requirements.txt
Jinja2==3.0.3
pandas==1.1.5
xlrd==1.2.0
python-dotenv==0.20.0
